foma.el: Support serif fonts
* foma.el: Support serif fonts. Users can specify the serif fonts to
use in font profiles. Whether to serif fonts depends on an option
specified in font profiles, if that is missing, there is a package
global default option, default to nil, that is, not using serif
fonts. Also, fixed pitch font is not mandatory in font
profiles. Introduced breaking changes: Renaming properties of font
profiles.